+++ src/app/community/(postUpdate)/post/new/page.tsx
@@ -1,30 +1,22 @@
 'use client';
 import dynamic from 'next/dynamic';
 import styles from '@/styles/community/post/postNewPage.module.css';
-import { useRecoilValue } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { editorText } from '@/recoil/atom/editorAtom';
 import { useState } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { supabase } from '@/utils/supabase/supabase';
 import { communityPostData } from '@/apis/communityPostData';
 import BookSearch from '@/components/community/post/BookSearch';
-import { book_id, book_name } from '@/recoil/atom/\bbookIdAtom';
+import { book_id, book_name, selectBookData } from '@/recoil/atom/\bbookIdAtom';
 import { useInputState } from '@/hooks/useInputState';
-const Select = dynamic(() => import('react-select'), {
-	ssr: false,
-	loading: () => (
-		<div
-			style={{
-				width: '500px',
-				height: '38px',
-				backgroundColor: '#a5a5a5',
-				borderRadius: '5px',
-				marginRight: '50px',
-			}}></div>
-	),
-});
+import OptionBookReport from '@/components/community/post/OptionBookReport';
+import OptionBookMeeting from '@/components/community/post/OptionBookMeeting';
+import OptionBookBuying from '@/components/community/post/OptionBookBuying';
+import OptionBookSelling from '@/components/community/post/OptionBookSelling';
+
 const Editor = dynamic(
-	() => import('@/components/community/post/WysiwygEditor'),
+	() => import('@/components/community/common/WysiwygEditor'),
 	{
 		ssr: false,
 		loading: () => (
@@ -56,7 +48,7 @@ export default function postNewPage() {
 	const sellingPrice = useInputState('');
 	// select book id
 	const book_Id = useRecoilValue(book_id);
-	const bookName = useRecoilValue(book_name);
+	const seletedBook = useSetRecoilState(selectBookData);
 	// select state & onchange event
 	const [recruitmentNumber, setRecruitmentNumber] = useState<number>(0);
 	const onchangeRecruitmentNumber = (e: any) => {
@@ -112,135 +104,51 @@ export default function postNewPage() {
 	};
 	const pageSelectArea = () => {
 		if (page === 'bookReport') {
-			return (
-				<div className={styles.reportSelectContainer}>
-					<div className={styles.reportSelectWrap}>
-						<label>책을 선택해주세요.</label>
-						<BookSearch />
-					</div>
-					<div className={styles.reportSelectWrap}>
-						<label>선택한 책</label>
-						<div className={styles.selectBookText}>{bookName}</div>
-					</div>
-				</div>
-			);
+			return <OptionBookReport />;
 		}
 		if (page === 'bookMeeting') {
-			const recruitmentOptions = [
-				{ value: '', label: '모집 인원을 선택해주세요.' },
-				{ value: '1', label: '1명' },
-				{ value: '2', label: '2명' },
-				{ value: '3', label: '3명' },
-				{ value: '4', label: '4명' },
-				{ value: '5', label: '5명' },
-				{ value: '6', label: '6명' },
-				{ value: '7', label: '7명' },
-				{ value: '8', label: '8명' },
-				{ value: '9', label: '9명' },
-				{ value: '10', label: '10명 이상' },
-			];
 			return (
-				<div className={styles.meetingSelectContainer}>
-					<div className={styles.meetingSelectWrap}>
-						<label>연락 방법</label>
-						<input
-							type="text"
-							placeholder="카카오 오픈채팅방 URL을 입력해주세요."
-							value={(chatUrl.value as string) || ''}
-							onChange={chatUrl.onChange}
-						/>
-					</div>
-					<div className={styles.meetingSelectWrap}>
-						<label>모집 마감일</label>
-						<input
-							type="date"
-							className={styles.dateInput}
-							onChange={deadline.onChange}
-						/>
-					</div>
-					<div className={styles.meetingSelectWrap}>
-						<label>모집 인원</label>
-						<Select
-							className={styles.bookSelectBtn}
-							options={recruitmentOptions}
-							defaultValue={recruitmentOptions[0]}
-							onChange={onchangeRecruitmentNumber}
-						/>
-					</div>
-				</div>
+				<OptionBookMeeting
+					chatUrl={
+						chatUrl as {
+							value: string;
+							onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+						}
+					}
+					deadline={
+						deadline as {
+							value: Date;
+							onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+						}
+					}
+					onchangeRecruitmentNumber={onchangeRecruitmentNumber}
+				/>
 			);
 		}
 		if (page === 'bookBuying') {
 			return (
-				<div className={styles.buyingSelectContainer}>
-					<div className={styles.buyingSelectWrap}>
-						<label>책을 선택해주세요.</label>
-						<Select className={styles.bookSelectBtn} />
-					</div>
-					<div className={styles.buyingSelectWrap}>
-						<label>선택한 책</label>
-						<div className={styles.selectBookText}></div>
-					</div>
-					<div className={styles.buyingSelectWrap}>
-						<label>구매하고 싶은 가격을 적어주세요.</label>
-						<input
-							type="text"
-							value={sellingPrice.value as string}
-							onChange={sellingPrice.onChange}
-						/>
-					</div>
-				</div>
+				<OptionBookBuying
+					sellingPrice={
+						sellingPrice as {
+							value: string;
+							onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+						}
+					}
+				/>
 			);
 		}
 		if (page === 'bookSelling') {
-			const bookStateOptions = [
-				{ value: '', label: '책의 상태를 선택해주세요.' },
-				{ value: '상', label: '상' },
-				{ value: '중', label: '중' },
-				{ value: '하', label: '하' },
-			];
-			const sellingStateOptions = [
-				{ value: '', label: '판매 / 나눔' },
-				{ value: '판매', label: '판매' },
-				{ value: '나눔', label: '나눔' },
-			];
 			return (
-				<div className={styles.sellingSelectContainer}>
-					<div className={styles.sellingSelectWrap}>
-						<label>책을 선택해주세요.</label>
-						<Select className={styles.bookSelectBtn} />
-					</div>
-					<div className={styles.sellingSelectWrap}>
-						<label>선택한 책</label>
-						<div className={styles.selectBookText}></div>
-					</div>
-					<div className={styles.sellingSelectWrap}>
-						<label>판매하고 싶은 가격을 적어주세요.</label>
-						<input
-							type="text"
-							value={sellingPrice.value as string}
-							onChange={sellingPrice.onChange}
-						/>
-					</div>
-					<div className={styles.sellingSelectWrap}>
-						<label>책의 상태</label>
-						<Select
-							className={styles.bookSelectBtn}
-							options={bookStateOptions}
-							defaultValue={bookStateOptions[0]}
-							onChange={onchangeBookState}
-						/>
-					</div>
-					<div className={styles.sellingSelectWrap}>
-						<label>판매 / 나눔</label>
-						<Select
-							className={styles.bookSelectBtn}
-							options={sellingStateOptions}
-							defaultValue={sellingStateOptions[0]}
-							onChange={onchangeSellingState}
-						/>
-					</div>
-				</div>
+				<OptionBookSelling
+					sellingPrice={
+						sellingPrice as {
+							value: string;
+							onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+						}
+					}
+					onchangeBookState={onchangeBookState}
+					onchangeSellingState={onchangeSellingState}
+				/>
 			);
 		}
 	};
