+++ src/styles/community/post/OptionBookSelling.module.css
@@ -1,0 +1,38 @@
+/* 팝니다 옵션 area css */
+.sellingSelectContainer {
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+	row-gap: 40px;
+	margin-bottom: 30px;
+}
+.sellingSelectWrap {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+}
+.sellingSelectWrap > label {
+	font-size: 24px;
+	font-weight: bold;
+	margin-bottom: 10px;
+}
+.sellingSelectWrap > input {
+	width: 478px;
+	height: 38px;
+	font-size: 16px;
+	padding: 0 10px;
+	border: 1px solid #cccccc;
+	border-radius: 5px;
+}
+.selectBookText {
+	width: 478px;
+	height: 38px;
+	font-size: 16px;
+	padding: 0 10px;
+	border: 1px solid #cccccc;
+	border-radius: 5px;
+	line-height: 38px;
+}
+.bookSelectBtn {
+	width: 500px;
+	margin-right: 50px;
+}
